backend/requirements.txt
tensorflow
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy